Rubrics: 1. change points from Integer to double; 2. add more optional fields 3. Rating -> Ratings

+++ src/main/java/edu/ksu/canvas/model/assignment/Submission.java
@@ -229,7 +229,7 @@ public class Submission extends BaseCanvasModel implements Serializable {
 
 	    private String ratingId;
 	    private String comments;
-	    private Integer points;
+	    private Double points;
 	    
 	    public String getRatingId() {
 	        return ratingId;
@@ -247,11 +247,11 @@ public class Submission extends BaseCanvasModel implements Serializable {
 	        this.comments = comments;
 	    }
 	    
-	    public Integer getPoints() {
+	    public Double getPoints() {
 	 		return points;
 	 	}
 
-	 	public void setPoints(Integer points) {
+	 	public void setPoints(Double points) {
 	 		this.points = points;
 	 	}
 	}
